+++ Board_LEDs.h
@@ -1,0 +1,3 @@
+void board_LED_init();
+
+void change_color(int switch_num);
