+++ Defines.h
@@ -1,0 +1,16 @@
+#define red_led		(18)		// PORT B
+#define green_led	(19)		// PORT B
+#define blue_led	(1)			// PORT D
+
+#define led_1	(3)					// PORT E
+#define led_2	(4)					// PORT E
+#define led_3	(5)					// PORT E
+
+#define switch_1	(8)			// PORT C
+#define switch_2	(9)			// PORT C
+#define switch_3	(10)		// PORT C
+#define switch_4	(11)		// PORT C
+#define switch_5	(12)		// PORT C
+#define switch_6	(13)		// PORT C
+
+#define MASK(x) (1UL << (x))
