+++ My_LEDs.h
@@ -1,0 +1,5 @@
+#include "stdbool.h"										// Booleans
+
+void LED_init();
+
+void set_LEDs(bool bit_1, bool bit_2, bool bit_3);
