+++ Objects/Digital_filter.build_log.htm
@@ -21,20 +21,27 @@ Target DLL:      CMSIS_AGDI.dll V1.28.3.0
 Dialog DLL:      TARMCM1.DLL V1.13.0.0
  
 <h2>Project:</h2>
-C:\Code\Digital_Filter1\Digital_filter.uvprojx
-Project File Date:  03/14/2018
+C:\Code\Digital_Filter\Digital_filter.uvprojx
+Project File Date:  03/15/2018
 
 <h2>Output:</h2>
 *** Using Compiler 'V5.06 update 5 (build 528)', folder: 'C:\App\Kiel\ARM\ARMCC\Bin'
 Build target 'ULink2'
-assembling startup_MKL25Z4.s...
-compiling RTX_Conf_CM.c...
-compiling LED_FRDM-KL25Z.c...
-compiling system_MKL25Z4.c...
-compiling main.c...
+compiling My_LEDs.c...
+My_LEDs.h(3): warning:  #1295-D: Deprecated declaration LED_init - give arg types
+  void LED_init();
+My_LEDs.c: 1 warning, 0 errors
+compiling Board_LEDs.c...
+Board_LEDs.h(1): warning:  #1295-D: Deprecated declaration board_LED_init - give arg types
+  void board_LED_init();
+Board_LEDs.c: 1 warning, 0 errors
+compiling Switches.c...
+Switches.h(1): warning:  #1295-D: Deprecated declaration switch_init - give arg types
+  void switch_init();
+Switches.c: 1 warning, 0 errors
 linking...
 Program Size: Code=6316 RO-data=328 RW-data=56 ZI-data=3472  
-".\Objects\Digital_filter.axf" - 0 Error(s), 0 Warning(s).
+".\Objects\Digital_filter.axf" - 0 Error(s), 3 Warning(s).
 
 <h2>Software Packages used:</h2>
 
@@ -71,26 +78,26 @@ Package Vendor: Keil
 
    * Component: Keil::Device:Startup:2.5.0
       Source file:   Device\Source\ARM\startup_MKL25Z4.s
-      Source file:   Device\Source\system_MKL25Z4.c
       Include file:  Device\Source\system_MKL25Z4.h
+      Source file:   Device\Source\system_MKL25Z4.c
 
    * Component: Keil.FRDM-KL25Z::Board Support:LED:1.0.0
       Source file:   Boards\NXP\FRDM-KL25Z\Common\LED_FRDM-KL25Z.c
 
    * Component: ARM::CMSIS:RTOS:Keil RTX:4.81.1
-      Source file:   CMSIS\RTOS\RTX\UserCodeTemplates\main.c
-      Library file:  CMSIS\RTOS\RTX\LIB\ARM\RTX_CM0.lib
       Include file:  CMSIS\RTOS\RTX\UserCodeTemplates\osObjects.h
-      Include file:  CMSIS\RTOS\RTX\INC\cmsis_os.h
-      Source file:   CMSIS\RTOS\RTX\UserCodeTemplates\MemPool.c
-      Source file:   CMSIS\RTOS\RTX\UserCodeTemplates\MsgQueue.c
-      Source file:   CMSIS\RTOS\RTX\UserCodeTemplates\Thread.c
-      Source file:   CMSIS\RTOS\RTX\Templates\RTX_Conf_CM.c
       Source file:   CMSIS\RTOS\RTX\UserCodeTemplates\MailQueue.c
+      Source file:   CMSIS\RTOS\RTX\UserCodeTemplates\MsgQueue.c
+      Source file:   CMSIS\RTOS\RTX\UserCodeTemplates\Mutex.c
+      Source file:   CMSIS\RTOS\RTX\Templates\RTX_Conf_CM.c
+      Source file:   CMSIS\RTOS\RTX\SRC\ARM\SVC_Table.s
+      Source file:   CMSIS\RTOS\RTX\UserCodeTemplates\MemPool.c
+      Include file:  CMSIS\RTOS\RTX\INC\cmsis_os.h
+      Source file:   CMSIS\RTOS\RTX\UserCodeTemplates\Thread.c
+      Source file:   CMSIS\RTOS\RTX\UserCodeTemplates\main.c
       Source file:   CMSIS\RTOS\RTX\UserCodeTemplates\Semaphore.c
       Source file:   CMSIS\RTOS\RTX\UserCodeTemplates\Timer.c
-      Source file:   CMSIS\RTOS\RTX\SRC\ARM\SVC_Table.s
-      Source file:   CMSIS\RTOS\RTX\UserCodeTemplates\Mutex.c
+      Library file:  CMSIS\RTOS\RTX\LIB\ARM\RTX_CM0.lib
 
    * Component: ARM::CMSIS:CORE:5.1.1
 
@@ -98,7 +105,7 @@ Package Vendor: Keil
 
    * Component: ::Board Support:LED:1.0.0 (API)
       Include file:  Board\Board_LED.h
-Build Time Elapsed:  00:00:01
+Build Time Elapsed:  00:00:00
 </pre>
 </body>
 </html>
