+++ Objects/Digital_filter.htm
@@ -3,7 +3,7 @@
 <title>Static Call Graph - [.\Objects\Digital_filter.axf]</title></head>
 <body><HR>
 <H1>Static Call Graph for image .\Objects\Digital_filter.axf</H1><HR>
-<BR><P>#&#060CALLGRAPH&#062# ARM Linker, 5060528: Last Updated: Wed Mar 14 20:52:11 2018
+<BR><P>#&#060CALLGRAPH&#062# ARM Linker, 5060528: Last Updated: Thu Mar 15 20:44:35 2018
 <BR><P>
 <H3>Maximum Stack Usage =        108 bytes + Unknown(Functions without stacksize, Cycles, Untraceable Function Pointers)</H3><H3>
 Call chain for Maximum Stack Depth:</H3>
@@ -437,7 +437,7 @@ Global Symbols
 
 <P><STRONG><a name="[c2]"></a>__semihosting_library_function</STRONG> (Thumb, 0 bytes, Stack size unknown bytes, indicate_semi.o(.text), UNUSED)
 
-<P><STRONG><a name="[62]"></a>LED_init</STRONG> (Thumb, 64 bytes, Stack size 0 bytes, main.o(i.LED_init))
+<P><STRONG><a name="[62]"></a>LED_init</STRONG> (Thumb, 64 bytes, Stack size 0 bytes, my_leds.o(i.LED_init))
 <BR><BR>[Called By]<UL><LI><a href="#[61]">&gt;&gt;</a>&nbsp;&nbsp;&nbsp;initialize_board
 </UL>
 
@@ -480,12 +480,12 @@ Global Symbols
 <BR><BR>[Stack]<UL><LI>Max Depth = 8<LI>Call Chain = _mutex_release
 </UL>
 
-<P><STRONG><a name="[63]"></a>board_LED_init</STRONG> (Thumb, 154 bytes, Stack size 0 bytes, main.o(i.board_LED_init))
+<P><STRONG><a name="[63]"></a>board_LED_init</STRONG> (Thumb, 154 bytes, Stack size 0 bytes, board_leds.o(i.board_LED_init))
 <BR><BR>[Called By]<UL><LI><a href="#[61]">&gt;&gt;</a>&nbsp;&nbsp;&nbsp;initialize_board
 </UL>
 
-<P><STRONG><a name="[61]"></a>initialize_board</STRONG> (Thumb, 30 bytes, Stack size 4 bytes, main.o(i.initialize_board))
-<BR><BR>[Stack]<UL><LI>Max Depth = 4<LI>Call Chain = initialize_board
+<P><STRONG><a name="[61]"></a>initialize_board</STRONG> (Thumb, 30 bytes, Stack size 8 bytes, main.o(i.initialize_board))
+<BR><BR>[Stack]<UL><LI>Max Depth = 8<LI>Call Chain = initialize_board
 </UL>
 <BR>[Calls]<UL><LI><a href="#[64]">&gt;&gt;</a>&nbsp;&nbsp;&nbsp;switch_init
 <LI><a href="#[63]">&gt;&gt;</a>&nbsp;&nbsp;&nbsp;board_LED_init
@@ -947,7 +947,7 @@ Global Symbols
 <BR><BR>[Called By]<UL><LI><a href="#[37]">&gt;&gt;</a>&nbsp;&nbsp;&nbsp;svcThreadGetId
 </UL>
 
-<P><STRONG><a name="[69]"></a>set_LEDs</STRONG> (Thumb, 94 bytes, Stack size 8 bytes, main.o(i.set_LEDs))
+<P><STRONG><a name="[69]"></a>set_LEDs</STRONG> (Thumb, 94 bytes, Stack size 8 bytes, my_leds.o(i.set_LEDs))
 <BR><BR>[Stack]<UL><LI>Max Depth = 8<LI>Call Chain = set_LEDs
 </UL>
 <BR>[Called By]<UL><LI><a href="#[3d]">&gt;&gt;</a>&nbsp;&nbsp;&nbsp;main
@@ -1024,7 +1024,7 @@ Global Symbols
 </UL>
 <BR>[Address Reference Count : 1]<UL><LI> rt_cmsis.o(i.osTimerCall)
 </UL>
-<P><STRONG><a name="[64]"></a>switch_init</STRONG> (Thumb, 60 bytes, Stack size 0 bytes, main.o(i.switch_init))
+<P><STRONG><a name="[64]"></a>switch_init</STRONG> (Thumb, 60 bytes, Stack size 0 bytes, switches.o(i.switch_init))
 <BR><BR>[Called By]<UL><LI><a href="#[61]">&gt;&gt;</a>&nbsp;&nbsp;&nbsp;initialize_board
 </UL>
