+++ Objects/Digital_filter.lnp
@@ -1,5 +1,9 @@
 --cpu Cortex-M0+
 ".\objects\main.o"
+".\objects\board_leds.o"
+".\objects\display.o"
+".\objects\my_leds.o"
+".\objects\switches.o"
 ".\objects\led_frdm-kl25z.o"
 "C:\App\Kiel\ARM\PACK\ARM\CMSIS\5.3.0\CMSIS\RTOS\RTX\LIB\ARM\RTX_CM0.lib"
 ".\objects\rtx_conf_cm.o"
