+++ Objects/board_leds.d
@@ -1,0 +1,10 @@
+.\objects\board_leds.o: Board_LEDs.c
+.\objects\board_leds.o: C:\App\Kiel\ARM\PACK\Keil\Kinetis_KLxx_DFP\1.14.0\Device\Include\MKL25Z4.h
+.\objects\board_leds.o: C:\App\Kiel\ARM\ARMCC\Bin\..\include\stdint.h
+.\objects\board_leds.o: C:\App\Kiel\ARM\PACK\ARM\CMSIS\5.3.0\CMSIS\Include\core_cm0plus.h
+.\objects\board_leds.o: C:\App\Kiel\ARM\PACK\ARM\CMSIS\5.3.0\CMSIS\Include\cmsis_version.h
+.\objects\board_leds.o: C:\App\Kiel\ARM\PACK\ARM\CMSIS\5.3.0\CMSIS\Include\cmsis_compiler.h
+.\objects\board_leds.o: C:\App\Kiel\ARM\PACK\ARM\CMSIS\5.3.0\CMSIS\Include\cmsis_armcc.h
+.\objects\board_leds.o: .\RTE\Device\MKL25Z128xxx4\system_MKL25Z4.h
+.\objects\board_leds.o: Defines.h
+.\objects\board_leds.o: Board_LEDs.h
