+++ Objects/display.d
@@ -1,0 +1,10 @@
+.\objects\display.o: Display.c
+.\objects\display.o: C:\App\Kiel\ARM\PACK\Keil\Kinetis_KLxx_DFP\1.14.0\Device\Include\MKL25Z4.h
+.\objects\display.o: C:\App\Kiel\ARM\ARMCC\Bin\..\include\stdint.h
+.\objects\display.o: C:\App\Kiel\ARM\PACK\ARM\CMSIS\5.3.0\CMSIS\Include\core_cm0plus.h
+.\objects\display.o: C:\App\Kiel\ARM\PACK\ARM\CMSIS\5.3.0\CMSIS\Include\cmsis_version.h
+.\objects\display.o: C:\App\Kiel\ARM\PACK\ARM\CMSIS\5.3.0\CMSIS\Include\cmsis_compiler.h
+.\objects\display.o: C:\App\Kiel\ARM\PACK\ARM\CMSIS\5.3.0\CMSIS\Include\cmsis_armcc.h
+.\objects\display.o: .\RTE\Device\MKL25Z128xxx4\system_MKL25Z4.h
+.\objects\display.o: Defines.h
+.\objects\display.o: Display.h
