+++ Objects/switches.d
@@ -1,0 +1,10 @@
+.\objects\switches.o: Switches.c
+.\objects\switches.o: C:\App\Kiel\ARM\PACK\Keil\Kinetis_KLxx_DFP\1.14.0\Device\Include\MKL25Z4.h
+.\objects\switches.o: C:\App\Kiel\ARM\ARMCC\Bin\..\include\stdint.h
+.\objects\switches.o: C:\App\Kiel\ARM\PACK\ARM\CMSIS\5.3.0\CMSIS\Include\core_cm0plus.h
+.\objects\switches.o: C:\App\Kiel\ARM\PACK\ARM\CMSIS\5.3.0\CMSIS\Include\cmsis_version.h
+.\objects\switches.o: C:\App\Kiel\ARM\PACK\ARM\CMSIS\5.3.0\CMSIS\Include\cmsis_compiler.h
+.\objects\switches.o: C:\App\Kiel\ARM\PACK\ARM\CMSIS\5.3.0\CMSIS\Include\cmsis_armcc.h
+.\objects\switches.o: .\RTE\Device\MKL25Z128xxx4\system_MKL25Z4.h
+.\objects\switches.o: Defines.h
+.\objects\switches.o: Switches.h
